refactor: allow messages to be printed through lint-staged
