+++ .lint-stagedrc.js
@@ -3,7 +3,7 @@ const stage = process.env.CLI_STYLE_STAGE === 'true'
 
 module.exports = {
     '*.{js,jsx,ts,tsx}': [
-        `d2-style js ${fix ? 'apply' : 'check'} ${
+        `./bin/d2-style js ${fix ? 'apply' : 'check'} ${
             fix && stage ? '--stage' : ''
         }`,
     ],
