+++ package.json
@@ -34,7 +34,7 @@
   "husky": {
     "hooks": {
       "commit-msg": "./bin/d2-style commit check",
-      "pre-commit": "./bin/d2-style validate js package --fix"
+      "pre-commit": "./bin/d2-style validate --lint-staged-config .lint-stagedrc.js"
     }
   },
   "publishConfig": {
