+++ src/cmds/js_cmds/apply.js
@@ -6,7 +6,7 @@ const { selectFiles } = require('../../files.js')
 const { stageFiles } = require('../../git-files.js')
 
 const { runner } = require('../../tools/js')
-const { PRETTIER_CONFIG, ESLINT_CONFIG } = require('../../config.js')
+const { PRETTIER_CONFIG, ESLINT_CONFIG } = require('../../paths.js')
 
 exports.command = 'apply [files..]'
 
@@ -25,29 +25,13 @@ exports.builder = {
         type: 'boolean',
         default: 'false',
     },
-    eslintConfig: {
-        describe: 'Override the ESLint configuration.',
-        type: 'string',
-        default: ESLINT_CONFIG,
-    },
-    prettierConfig: {
-        describe: 'Override the Prettier configuration.',
-        type: 'string',
-        default: PRETTIER_CONFIG,
-    },
 }
 
 exports.handler = argv => {
-    const { all, stage, files, eslintConfig, prettierConfig } = argv
+    const { all, stage, files } = argv
 
     const root = process.cwd()
     log.debug(`Root directory: ${root}`)
-
-    process.env = {
-        ...process.env,
-        CLI_STYLE_ESLINT_CONFIG: eslintConfig,
-        CLI_STYLE_PRETTIER_CONFIG: prettierConfig,
-    }
 
     const codeFiles = selectFiles(files, all, root)
     const report = runner(codeFiles, true)
@@ -55,7 +39,7 @@ exports.handler = argv => {
     report.summarize()
 
     if (report.hasViolations) {
-        log.error(
+        log.print(
             `${report.violations.length} file(s) violate the code standard.`
         )
         process.exit(1)
