+++ src/cmds/js_cmds/check.js
@@ -2,7 +2,7 @@ const log = require('@dhis2/cli-helpers-engine').reporter
 
 const { runner } = require('../../tools/js')
 const { selectFiles } = require('../../files.js')
-const { PRETTIER_CONFIG, ESLINT_CONFIG } = require('../../config.js')
+const { PRETTIER_CONFIG, ESLINT_CONFIG } = require('../../paths.js')
 
 exports.command = 'check [files..]'
 
@@ -15,29 +15,13 @@ exports.builder = {
         type: 'boolean',
         default: 'false',
     },
-    eslintConfig: {
-        describe: 'Override the ESLint configuration.',
-        type: 'string',
-        default: ESLINT_CONFIG,
-    },
-    prettierConfig: {
-        describe: 'Override the Prettier configuration.',
-        type: 'string',
-        default: PRETTIER_CONFIG,
-    },
 }
 
 exports.handler = argv => {
-    const { files, all, eslintConfig, prettierConfig } = argv
+    const { files, all } = argv
 
     const root = process.cwd()
     log.debug(`Root directory: ${root}`)
-
-    process.env = {
-        ...process.env,
-        CLI_STYLE_ESLINT_CONFIG: eslintConfig,
-        CLI_STYLE_PRETTIER_CONFIG: prettierConfig,
-    }
 
     const codeFiles = selectFiles(files, all, root)
     const report = runner(codeFiles)
@@ -45,7 +29,7 @@ exports.handler = argv => {
     report.summarize()
 
     if (report.hasViolations) {
-        log.error(
+        log.print(
             `${report.violations.length} file(s) violate the code standard.`
         )
         process.exit(1)
